Try to speed up

+++ train.py
@@ -59,7 +59,7 @@ if __name__ == "__main__":
         accelerator="gpu" if torch.cuda.is_available() else "cpu",
         devices=args.gpus,
         max_epochs=args.epochs,
-        precision=32,
+        precision="bf16-mixed",
         log_every_n_steps=10,
         val_check_interval=1.0,
         callbacks=[
